test(base),feat(Compiler),fix(base):提供\b\a\r\f\v\0等多个新换行符适配，同时将"include方式改为<>include方式，修复autodll无名字报错问题，增加auto.windows适配,增加autodll.linux适配，修复autodll.linux.dllopen漏洞

+++ src/include/base/utils/bytes.h
@@ -1,6 +1,6 @@
 ﻿#pragma once
 
-#include "base/trcdef.h"
+#include <base/trcdef.h>
 #include <cstdint>
 
 namespace trc::utils {
@@ -12,7 +12,7 @@ namespace trc::utils {
  * char正好是一个字节，而且比较好操作，使用之前需要把指针转为（char*）
  */
 TRC_base_func_api void bytes_order_change(
-    def::byte_t* const& bytes_, int size);
+    def::byte_t* const& bytes_, size_t size);
 
 uint32_t utf_8_decode();
